+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `get_url`, `unarchive` support for `owner`, `group`, `mode`, `setype`, `selevel`
 * `wait_for` module `port` parameters validation fix - integer
 * `firewalld` permanent type fix
+* `pause` validation fix for `minutes` and `seconds`
 
 # 1.0.6 - 14 October 2016
 * Allow sharing variables between tasks (#29)